refactor(client): reduce imports in routines feature internals
The routines edit/view/tracker route files had grown well past the
import/max-dependencies threshold (10). Decompose them following the
established "componentize first; barrel where cohesive" pattern so each
file pulls from fewer modules and the responsibilities are clearer.

- Add group barrels (layout, routines; extend dailies) to collapse
  several same-group imports into one.
- Extract bundled hooks (useRoutineTracker, useRoutineStatusMutation,
  useRoutineDetailsForm) that own the queries/mutations/derived state.
- Extract route-private components (tracker tables, routine today card
  + details content, new-routine form, shared quick-fill menu).
- Add makeDuplicateHandler to useEditFormPage (parallel to
  makeDeleteHandler) so the edit route drops its own sonner import.

All five target files are now <=10 non-type imports. Also dropped a
stale code-duplication suppression and trimmed speculative barrel
re-exports to avoid introducing dead code.
