+++ packages/client/src/components/dailies/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/max-dependencies -- cohesive feature barrel (precedent: utils/index.ts) */
 export { DailiesActiveListView } from "./DailiesActiveListView";
 export { DailyCriteriaTemplateEditModal } from "./DailyCriteriaTemplateEditModal";
 export { DailiesViewModeToggle } from "./DailiesViewModeToggle";
@@ -7,3 +8,11 @@ export { DailyTitle } from "./DailyTitle";
 export { TodayStatusCell } from "./TodayStatusCell";
 export { TooManyDailiesWarning } from "./TooManyDailiesWarning";
 export { DailiesLimitSetting } from "./DailiesLimitSetting";
+export { DailyDetailsPanel } from "./DailyDetailsPanel";
+export { DailyTrackerRow } from "./DailyTrackerRow";
+export { buildDailyTrackerColumns } from "./dailyTrackerColumns";
+export {
+  DAILY_DETAIL_TABS,
+  DAILY_STATUS_OPTIONS,
+  type DailyDetailTab,
+} from "./dailyStatusMeta";
