+++ packages/client/src/components/layout/index.ts
@@ -1,2 +1,10 @@
+// Barrel for the page-chrome / layout components so a page that renders several
+// of them imports from one module. Layout components import their siblings via
+// direct file paths (never this index), so this re-export introduces no cycle.
 export { DropdownNavItem } from "./DropdownNavItem";
+export { EditForm } from "./EditForm";
+export { EditPageFooter } from "./EditPageFooter";
+export { InfoArea } from "./InfoArea";
 export { NavDropdown } from "./NavDropdown";
+export { PageHeader } from "./PageHeader";
+export { PageTabs } from "./PageTabs";
