+++ packages/client/src/components/routines/index.ts
@@ -1,0 +1,15 @@
+// Barrel for the routine weekly-schedule helpers + editor components so a page
+// that uses several of them imports from one module. These modules import their
+// siblings via direct file paths (never this index), so there is no cycle.
+// Note: `weekly.test.ts` imports `./weekly` directly — leave it as-is.
+export {
+  DAY_LABELS,
+  DAY_ORDER,
+  fillAllDays,
+  representativeRow,
+  rowsToWeekly,
+  weeklyToRows,
+} from "./weekly";
+export { RoutineEntryLabel } from "./RoutineEntryLabel";
+export { WeeklyEntryEditor } from "./WeeklyEntryEditor";
+export { WeeklyScheduleField } from "./WeeklyScheduleField";
