+++ packages/client/src/routes/routines.$id.edit.-components/-DetailsTab.tsx
@@ -1,43 +1,19 @@
-import type { Routine } from "@emstack/types";
+import type { Routine, RoutineTemplate } from "@emstack/types";
 
-import { useEffect, useMemo, useRef, useState } from "react";
+import { Loader2 } from "lucide-react";
 
-import { useStore } from "@tanstack/react-form";
-import { useQuery } from "@tanstack/react-query";
-import { ChevronDownIcon, Loader2, WandSparklesIcon } from "lucide-react";
-import { toast } from "sonner";
-import * as z from "zod";
+import { QuickFillMenu } from "./-QuickFillMenu";
 
-import { useAppForm } from "@/components/formFields";
-import { EditForm } from "@/components/layout/EditForm";
+import { EditForm } from "@/components/layout";
 import {
   fillAllDays,
   representativeRow,
-  rowsToWeekly,
   weeklyToRows,
-} from "@/components/routines/weekly";
-import { WeeklyEntryEditor } from "@/components/routines/WeeklyEntryEditor";
-import { WeeklyScheduleField } from "@/components/routines/WeeklyScheduleField";
+  WeeklyEntryEditor,
+  WeeklyScheduleField,
+} from "@/components/routines";
 import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import {
-  buildConnectionOptions,
-  decodeConnection,
-  encodeConnection,
-  fetchResources,
-  fetchRoutineTemplates,
-  fetchTasks,
-  fetchTopics,
-  formHasChanges,
-  toOptions,
-  upsertRoutine,
-} from "@/utils";
-import { queryKeys } from "@/utils/queryKeys";
+import { useRoutineDetailsForm } from "@/hooks/useRoutineDetailsForm";
 
 const STATUS_OPTIONS = [
   {
@@ -69,33 +45,6 @@ const MODE_OPTIONS = [
   },
 ];
 
-const weeklyRowSchema = z
-  .object({
-    day: z.enum(["0", "1", "2", "3", "4", "5", "6"]),
-    type: z.enum(["", "task", "resource", "freeform"]),
-    id: z.string(),
-    notes: z.string(),
-    location: z.string(),
-    prependText: z.string(),
-    appendText: z.string(),
-  })
-  .refine(row => row.type === "" || row.id.length > 0, {
-    message: "Required",
-    path: ["id"],
-  });
-
-const detailsSchema = z.object({
-  name: z.string().min(1, "Name is required").max(255),
-  description: z.string().max(2000),
-  connections: z.array(z.string()),
-  status: z.enum(["active", "inactive", "complete", "paused"]),
-  mode: z.enum(["weekly", "daily"]),
-  weekly: z.array(weeklyRowSchema).length(7),
-  // Daily-mode only: how many days a week the routine needs doing. Null = no
-  // target (every day).
-  weeklyTarget: z.number().int().min(1).max(7).nullable(),
-});
-
 interface DetailsTabProps {
   routine: Routine;
   onSaved: () => Promise<void>;
@@ -108,111 +57,19 @@ export function DetailsTab({
   onChangeStateChange,
 }: DetailsTabProps) {
   const {
-    data: topics,
-  } = useQuery({
-    queryKey: ["topics"],
-    queryFn: () => fetchTopics(),
-  });
-
-  const {
-    data: tasks,
-  } = useQuery({
-    queryKey: ["tasks"],
-    queryFn: () => fetchTasks(),
-  });
-
-  const {
-    data: resources,
-  } = useQuery({
-    queryKey: queryKeys.resources.list(),
-    queryFn: () => fetchResources(),
-  });
-
-  const {
-    data: routineTemplates,
-  } = useQuery({
-    queryKey: ["routineTemplates"],
-    queryFn: () => fetchRoutineTemplates(),
-  });
-
-  const taskOptions = useMemo(() => toOptions(tasks), [tasks]);
-  const resourceOptions = useMemo(() => toOptions(resources), [resources]);
-  const connectionOptions = useMemo(
-    () => buildConnectionOptions(topics, tasks, resources),
-    [topics, tasks, resources],
-  );
-
-  const startingValues = useMemo(
-    () => ({
-      name: routine.name ?? "",
-      description: routine.description ?? "",
-      connections: (routine.connections ?? []).map(encodeConnection),
-      status: routine.status ?? "active",
-      mode: routine.mode ?? "weekly",
-      weekly: weeklyToRows(routine.weekly),
-      weeklyTarget: routine.weeklyTarget ?? null,
-    }),
-    [routine],
-  );
-
-  const lastSavedRef = useRef(startingValues);
-  const [isSaving, setIsSaving] = useState(false);
-
-  const form = useAppForm({
-    defaultValues: startingValues,
-    validators: {
-      onSubmit: detailsSchema,
-    },
-    onSubmit: async ({
-      value,
-    }) => {
-      setIsSaving(true);
-      try {
-        const connections = value.connections
-          .map(decodeConnection)
-          .filter((c): c is NonNullable<typeof c> => c !== null);
-
-        await upsertRoutine(routine.id, {
-          name: value.name,
-          description: value.description || null,
-          connections,
-          status: value.status,
-          mode: value.mode,
-          // Daily mode mirrors the single chosen entry onto all 7 days so
-          // "today's item" resolves identically every day.
-          weekly:
-            value.mode === "daily"
-              ? rowsToWeekly(fillAllDays(representativeRow(value.weekly)))
-              : rowsToWeekly(value.weekly),
-          // The weekly target only applies to daily routines; clear it for
-          // weekly schedules.
-          weeklyTarget: value.mode === "daily" ? value.weeklyTarget : null,
-        });
-        lastSavedRef.current = value;
-        onChangeStateChange?.(false);
-        await onSaved();
-        toast.success("Details saved.");
-      }
-      catch {
-        toast.error("Failed to save details.");
-      }
-      finally {
-        setIsSaving(false);
-      }
-    },
-  });
-
-  const currentValues = useStore(form.store, state => ({
-    ...state.values,
-  }));
-  const hasChanges = formHasChanges(currentValues, lastSavedRef.current);
-  const isDaily = currentValues.mode === "daily";
-
-  useEffect(() => {
-    onChangeStateChange?.(hasChanges);
-  }, [hasChanges, onChangeStateChange]);
+    form,
+    connectionOptions,
+    taskOptions,
+    resourceOptions,
+    isDaily,
+    isSaving,
+  } = useRoutineDetailsForm(routine, onSaved, onChangeStateChange);
 
   return (
+    // The EditForm + "Routine Name" + "Type" header is intentionally mirrored
+    // by the create form (-NewRoutineForm) — the create/edit split keeps the
+    // two forms separate, so the small overlap is accepted rather than shared
+    // through a field-group abstraction.
     // fallow-ignore-next-line code-duplication
     <EditForm
       onSubmit={form.handleSubmit}
@@ -287,39 +144,11 @@ export function DetailsTab({
                   className="flex flex-wrap items-center justify-between gap-2"
                 >
                   <span className="text-2xl">Weekly Schedule</span>
-                  <DropdownMenu modal={false}>
-                    <DropdownMenuTrigger asChild>
-                      <Button
-                        type="button"
-                        variant="outline"
-                        size="sm"
-                      >
-                        <WandSparklesIcon className="size-4" />
-                        Quick Fill
-                        <ChevronDownIcon className="size-4" />
-                      </Button>
-                    </DropdownMenuTrigger>
-                    <DropdownMenuContent align="end">
-                      {(routineTemplates ?? []).length === 0
-                        ? (
-                          <DropdownMenuItem disabled>
-                            No templates — add one in Settings
-                          </DropdownMenuItem>
-                        )
-                        : (routineTemplates ?? []).map(template => (
-                          <DropdownMenuItem
-                            key={template.id}
-                            onSelect={() => {
-                              field.handleChange(
-                                weeklyToRows(template.weekly),
-                              );
-                            }}
-                          >
-                            {template.label}
-                          </DropdownMenuItem>
-                        ))}
-                    </DropdownMenuContent>
-                  </DropdownMenu>
+                  <QuickFillMenu<RoutineTemplate>
+                    kind="routine"
+                    onSelect={template =>
+                      field.handleChange(weeklyToRows(template.weekly))}
+                  />
                 </div>
                 <WeeklyScheduleField
                   value={field.state.value}
