+++ packages/client/src/routes/routines.$id.index.tsx
@@ -1,39 +1,21 @@
-import type { DailyDetailTab } from "@/components/dailies/dailyStatusMeta";
-import type { Daily, DailyCompletionStatus } from "@emstack/types";
+import type { DailyDetailTab } from "@/components/dailies";
 
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { createFileRoute, Link, useNavigate } from "@tanstack/react-router";
-import { EditIcon, FlameIcon, LaughIcon } from "lucide-react";
-import { toast } from "sonner";
+import { EditIcon } from "lucide-react";
 
-import { EntityLink } from "@/components/boxElements/EntityLink";
-import { DashboardCard } from "@/components/boxes/DashboardCard";
-import { TodayStatusCell } from "@/components/dailies";
-import { DailyDetailsPanel } from "@/components/dailies/DailyDetailsPanel";
-import { DAILY_DETAIL_TABS } from "@/components/dailies/dailyStatusMeta";
+import { RoutineDetailsContent } from "./routines.$id.-components/-RoutineDetailsContent";
+import { RoutineTodayCard } from "./routines.$id.-components/-RoutineTodayCard";
+
+import {
+  DAILY_DETAIL_TABS,
+
+  DailyDetailsPanel,
+} from "@/components/dailies";
 import { EntityError, EntityPending } from "@/components/EntityStates";
-import { InfoArea } from "@/components/layout/InfoArea";
-import { PageHeader } from "@/components/layout/PageHeader";
-import { RoutineEntryLabel } from "@/components/routines/RoutineEntryLabel";
-import {
-  DAY_LABELS,
-  DAY_ORDER,
-} from "@/components/routines/weekly";
+import { PageHeader } from "@/components/layout";
 import { Button } from "@/components/ui/button";
-import { useSettings } from "@/hooks/useSettings";
-import { useTaskResourceNames } from "@/hooks/useTaskResourceNames";
-import {
-  connectionEntityKind,
-  fetchSingleRoutine,
-  findStatusForDate,
-  getCurrentChain,
-  getTodayKey,
-  getTotalCompletedDays,
-  isWeeklyTargetMet,
-  upsertRoutine,
-  withCompletion,
-  withCompletionNote,
-} from "@/utils";
+import { fetchSingleRoutine } from "@/utils";
 
 export interface RoutineViewSearch {
   tab?: DailyDetailTab;
@@ -63,19 +45,12 @@ function RoutineError() {
   return <EntityError entity="routine" />;
 }
 
-// Pre-existing complexity hotspot (untested route component); suppressed so
-// unrelated edits inside it don't trip the audit gate. Refactor candidate.
-// fallow-ignore-next-line complexity
 function SingleRoutine() {
   const {
     id,
   } = Route.useParams();
 
   const navigate = useNavigate();
-  const queryClient = useQueryClient();
-  const {
-    settings,
-  } = useSettings();
   const search = Route.useSearch();
   const tab: DailyDetailTab = search.tab ?? "details";
 
@@ -99,61 +74,6 @@ function SingleRoutine() {
     queryFn: () => fetchSingleRoutine(id),
   });
 
-  const {
-    taskNames,
-    resourceNames,
-  } = useTaskResourceNames();
-
-  const todayDateKey = getTodayKey();
-
-  // Marks today's completion status (incomplete/touched/goal/exceeded/freeze)
-  // for this routine — the same daily-status concept the dashboard table edits.
-  // Sends a partial body (name + completions), so the upsert preserves the
-  // routine's weekly grid, connections, criteria and overall status.
-  const statusMutation = useMutation({
-    mutationFn: ({
-      daily,
-      status,
-      note,
-    }: {
-      daily: Daily;
-      status: DailyCompletionStatus;
-      note?: string | null;
-    }) => {
-      const withStatus = withCompletion(daily, todayDateKey, status);
-      const completions
-        = note === undefined
-          ? withStatus
-          : withCompletionNote(
-            {
-              ...daily,
-              completions: withStatus,
-            },
-            todayDateKey,
-            note,
-          );
-      return upsertRoutine(daily.id, {
-        name: daily.name,
-        completions,
-      });
-    },
-    onSuccess: async () => {
-      await queryClient.invalidateQueries({
-        queryKey: ["routine", id],
-      });
-      await queryClient.invalidateQueries({
-        queryKey: ["routines"],
-      });
-      await queryClient.invalidateQueries({
-        queryKey: ["dailies"],
-      });
-      toast.success("Status updated.");
-    },
-    onError: () => {
-      toast.error("Failed to update status.");
-    },
-  });
-
   if (isPending) {
     return <RoutinePending />;
   }
@@ -161,64 +81,6 @@ function SingleRoutine() {
   if (error || !data) {
     return <RoutineError />;
   }
-
-  const weekly = data.weekly ?? {};
-  const isDaily = data.mode === "daily";
-  const dailyEntry = isDaily
-    ? Object.values(weekly).find(Boolean) ?? null
-    : null;
-  // The same day of the week as today (JS getDay: "0" = Sunday … "6" = Saturday).
-  // Daily routines mirror their entry onto every day, so this resolves to the
-  // single daily item; weekly routines resolve to today's scheduled entry (if any).
-  const todayKey = String(new Date().getDay());
-  const todayEntry = isDaily
-    ? dailyEntry
-    : (weekly[todayKey as keyof typeof weekly] ?? null);
-  const completions = data.completions ?? [];
-  // Daily-mode routines with a met weekly target need nothing more today.
-  const weekTargetMet
-    = isDaily
-      && isWeeklyTargetMet(
-        {
-          completions,
-          weeklyTarget: data.weeklyTarget ?? null,
-        },
-        todayDateKey,
-        settings.weekTargetWindow,
-      );
-  const chain = getCurrentChain({
-    completions,
-  });
-  const total = getTotalCompletedDays({
-    completions,
-  });
-
-  // A Daily-shaped view of this routine for the shared today's-status modal,
-  // which only reads name / completions / criteria / description.
-  const dailyForStatus: Daily = {
-    id: data.id,
-    name: data.name,
-    description: data.description,
-    completions: data.completions ?? [],
-    criteria: data.criteria ?? null,
-    status: data.status,
-  };
-  const todayStatus = findStatusForDate(dailyForStatus, todayDateKey);
-  const statusControl = (
-    <div className="w-40">
-      <TodayStatusCell
-        daily={dailyForStatus}
-        currentStatus={todayStatus}
-        disabled={statusMutation.isPending}
-        onChange={(status, note) =>
-          statusMutation.mutate({
-            daily: dailyForStatus,
-            status,
-            note,
-          })}
-      />
-    </div>
-  );
 
   return (
     <div>
@@ -240,39 +102,7 @@ function SingleRoutine() {
         </Link>
       </PageHeader>
       <div className="container flex flex-col gap-12">
-        <DashboardCard
-          title="Today's Task"
-          action={statusControl}
-        >
-          {todayEntry
-            ? (
-              <>
-                <p className="text-lg font-medium">
-                  <RoutineEntryLabel
-                    entry={todayEntry}
-                    taskNames={taskNames}
-                    resourceNames={resourceNames}
-                    showMeta={false}
-                  />
-                </p>
-                {todayEntry.notes && (
-                  <p className="text-sm text-muted-foreground">
-                    {todayEntry.notes}
-                  </p>
-                )}
-              </>
-            )
-            : (
-              <p className="text-muted-foreground italic">
-                Nothing, take a break!
-              </p>
-            )}
-          {weekTargetMet && (
-            <p className="mt-2 text-sm text-muted-foreground italic">
-              Nothing required today — weekly goal met.
-            </p>
-          )}
-        </DashboardCard>
+        <RoutineTodayCard data={data} />
         <DailyDetailsPanel
           dailyId={id}
           tab={tab}
@@ -297,134 +127,7 @@ function SingleRoutine() {
               </Button>
             </Link>
           )}
-          detailsContent={(
-            <div className="flex flex-col gap-12">
-              <div
-                className="
-                  grid grid-cols-1 gap-12
-                  md:grid-cols-4
-                "
-              >
-                <InfoArea
-                  header="Type"
-                  condition={true}
-                >
-                  <span
-                    className="
-                      inline-flex items-center rounded-full px-2.5 py-0.5
-                      text-xs font-medium
-                    "
-                  >
-                    {isDaily ? "Daily Task" : "Weekly Schedule"}
-                  </span>
-                </InfoArea>
-                <InfoArea
-                  header="Status"
-                  condition={true}
-                >
-                  <span className="capitalize">{data.status ?? "active"}</span>
-                </InfoArea>
-                <div className="md:col-span-2">
-                  <InfoArea
-                    header="Stats"
-                    condition={true}
-                  >
-                    <div className="flex flex-row flex-wrap gap-6 text-sm">
-                      <span className="inline-flex items-center gap-1">
-                        <FlameIcon
-                          size={16}
-                          className={chain > 0
-                            ? "text-orange-600"
-                            : "text-muted-foreground"}
-                        />
-                        <strong>{chain}</strong>
-                        <span className="text-muted-foreground">day chain</span>
-                      </span>
-                      <span className="inline-flex items-center gap-1">
-                        <LaughIcon
-                          size={16}
-                          className={total > 0
-                            ? "text-emerald-600"
-                            : "text-muted-foreground"}
-                        />
-                        <strong>{total}</strong>
-                        <span className="text-muted-foreground">total days</span>
-                      </span>
-                    </div>
-                  </InfoArea>
-                </div>
-              </div>
-              <InfoArea
-                header="Connected To"
-                condition={(data.connections?.length ?? 0) > 0}
-              >
-                <ul className="flex flex-col gap-1">
-                  {data.connections?.map(c => (
-                    <li key={`${c.type}:${c.id}`}>
-                      <EntityLink
-                        entity={connectionEntityKind(c.type)}
-                        id={c.id}
-                        className={`
-                          font-bold text-blue-800
-                          hover:text-blue-600
-                          dark:text-blue-300
-                        `}
-                      >
-                        <span
-                          className="
-                            mr-2 text-xs text-muted-foreground uppercase
-                          "
-                        >
-                          {c.type}
-                        </span>
-                        {c.name ?? c.id}
-                      </EntityLink>
-                    </li>
-                  ))}
-                </ul>
-              </InfoArea>
-              {!isDaily && (
-                <InfoArea
-                  header="Weekly Schedule"
-                  condition={true}
-                >
-                  <ul className="flex flex-col gap-1">
-                    {DAY_ORDER.map((day) => {
-                      const entry = weekly[day];
-                      return (
-                        <li
-                          key={day}
-                          className="
-                            grid grid-cols-[120px_1fr] items-center gap-2
-                            border-b border-border/60 py-1
-                          "
-                        >
-                          <span className="text-sm font-medium">
-                            {DAY_LABELS[day]}
-                          </span>
-                          {entry
-                            ? (
-                              <RoutineEntryLabel
-                                entry={entry}
-                                taskNames={taskNames}
-                                resourceNames={resourceNames}
-                              />
-                            )
-                            : (
-                              <span
-                                className="text-sm text-muted-foreground italic"
-                              >
-                                Nothing scheduled
-                              </span>
-                            )}
-                        </li>
-                      );
-                    })}
-                  </ul>
-                </InfoArea>
-              )}
-            </div>
-          )}
+          detailsContent={<RoutineDetailsContent data={data} />}
         />
       </div>
     </div>
